pyproject.toml
[project]
name = "agent-memory-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.115.0",
	"jinja2>=3.1.4",
	"langchain>=0.3.0",
	"langgraph>=0.2.0",
	"langchain-openai>=0.2.0",
	"pydantic>=2.9.0",
	"pyyaml>=6.0.0",
	"python-dotenv>=1.0.1",
	"uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_memory_demo"]
